remove hooks.go
