+++ controllers/argocd/TOBEREMOVED.go
@@ -10,6 +10,7 @@ import (
 	"os"
 	"sort"
 	"strings"
+	"sync"
 	"text/template"
 	"time"
 
@@ -1093,4 +1094,31 @@ func (r *ReconcileArgoCD) namespaceResourceMapper(ctx context.Context, o client.
 	}
 
 	return result
+}
+
+var (
+	mutex sync.RWMutex
+	hooks = []Hook{}
+)
+
+// Hook changes resources as they are created or updated by the reconciler.
+type Hook func(*argoproj.ArgoCD, interface{}, string) error
+
+// Register adds a modifier for updating resources during reconciliation.
+func Register(h ...Hook) {
+	mutex.Lock()
+	defer mutex.Unlock()
+	hooks = append(hooks, h...)
+}
+
+// nolint:unparam
+func applyReconcilerHook(cr *argoproj.ArgoCD, i interface{}, hint string) error {
+	mutex.Lock()
+	defer mutex.Unlock()
+	for _, v := range hooks {
+		if err := v(cr, i, hint); err != nil {
+			return err
+		}
+	}
+	return nil
 }
